# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Api desenvolvida para projeto de avaliação de candidatos para empresa Ioasys",
    "version": "1.0.0",
    "title": "IMDB - Clone",
    "contact": {
      "email": "saviohenrique.lage@gmail.com"
    }
  },

  "servers": [{ "url": "http://localhost:3000" }],
  "paths": {
    "/authenticate": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Login no sistema",
        "description": "",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginAdmin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucess"
          },
          "400": {
            "description": "Invalid password"
          },
          "404": {
            "description": "user not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users": {
      "post": {
        "tags": ["Usuário"],
        "summary": "Criação de usuário",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucess"
          },
          "400": {
            "description": "User already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": ["Usuário"],
        "summary": "Edição de usuário",
        "description": "Ação realizada somente pelo usuário que realiza a requisição",
        "operationId": "editUser",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucess"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": ["Usuário"],
        "summary": "Exclusão logica de usuário",
        "description": "Ação realizada somente pelo usuário que realiza a requisição",
        "operationId": "deleteUser",

        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Sucess"
          },
          "400": {
            "description": "User already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/admin": {
      "post": {
        "tags": ["Administrador"],
        "summary": "Criação de usuário/admin",
        "operationId": "createUserAdmin",
        "description": "Apenas administradores podem criar outros administradores",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucess"
          },
          "400": {
            "description": "User already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/admin/{userId}": {
      "put": {
        "tags": ["Administrador"],
        "summary": "Edição de usuário/admin",
        "description": "Apenas administradores podem editar outros usuarios",
        "operationId": "editUserAdmin",

        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucess"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": ["Administrador"],
        "summary": "Exclusão logica de usuário",
        "description": "Apenas administradores podem excluir outros usuarios",
        "operationId": "deleteUserAdmin",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucess"
          },
          "400": {
            "description": "User already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/movies": {
      "post": {
        "tags": ["Filmes"],
        "summary": "Criação de filmes",
        "operationId": "createMovie",

        "description": "Apenas administradores podem criar filmes",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucess"
          },
          "400": {
            "description": "User already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": ["Filmes"],
        "summary": "Listagem de filmes com passagem de filtros opcionais",

        "description": "Apenas usuarios/administradores logados podem listar filmes",
        "operationId": "listMovies",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genres",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actors",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "director",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Sucess"
          },
          "400": {
            "description": "User already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/movies/{movieId}": {
      "get": {
        "tags": ["Filmes"],
        "summary": "Detalhe do filme",

        "description": "Apenas usuarios/administradores logados podem detalhar filmes",
        "operationId": "detailMovie",
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Sucess"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },

    "/movies/vote": {
      "post": {
        "tags": ["Filmes"],
        "summary": "Votação",

        "description": "Apenas usuarios logados podem votar",
        "operationId": "vote",
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Vote"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucess"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },

    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "director": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "releaseDate": {
            "type": "string"
          },
          "actors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },

          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Vote": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "string",
            "format": "uuid"
          },

          "vote": {
            "type": "integer",
            "enum": [0, 1, 2, 3, 4]
          }
        }
      },
      "LoginAdmin": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "example": "default_email@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        }
      },
      "LoginUser": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "example": "default_user@email.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}